Extract custom functions from storage entity account
Only getters belongs to the shared Account entity from storage
component. The rest are customised functions owned by the
chain module.
