+++ framework/src/modules/chain/components/storage/sql/accounts/create_dependent_record.sql
