+++ framework/src/modules/chain/components/storage/sql/accounts/delete.sql
