+++ framework/src/modules/chain/components/storage/sql/accounts/delete_dependent_record.sql
