+++ framework/src/modules/chain/components/storage/sql/accounts/sync_delegates_rank.sql
